+++ lib/Validator.js
@@ -19,8 +19,9 @@
 const check = require('check-types');
 
 const messages = require('./provider/messages');
-const NoArgumentError = require('./error/NoArgumentError');
-const InvalidArgumentError = require('./error/InvalidArgumentError');
+const constants = require('./provider/constants');
+
+const ArgumentError = require('./error/NoArgumentError');
 
 /**
  * Class representing a validator.
@@ -40,7 +41,9 @@ class Validator {
   static _checkTypeArgument(type) {
     // Check type of `type` argument, if provided.
     if (check.assigned(type) && check.not.string(type)) {
-      throw new InvalidArgumentError(`Type argument ${messages.IS_NOT_STRING}`);
+      const message = `Type argument ${messages.IS_NOT_STRING}`;
+      const code = constants.ARGUMENT_WRONG_TYPE
+      throw new ArgumentError(message, code);
     }
   }
