+++ lib/error/ArgumentError.js
@@ -1,11 +1,11 @@
 /**
- * @name Invalid argument error
+ * @name Argument error
  *
  * @description Error class for argument type errors.
  *
  * @example
- * 	const InvalidArgumentError = require('./InvalidArgumentError');
- * 	throw new InvalidArgumentError('Expected string');
+ * 	const ArgumentError = require('./ArgumentError');
+ * 	throw new ArgumentError('Expected string', 2);
  */
 
 'use strict';
@@ -13,17 +13,28 @@
 const ExtendableError = require('./ExtendableError');
 
 /**
- * Class representing an invalid error class.
+ * Class representing an argument error.
  * @extends ExtendableError
  */
-class InvalidArgumentError extends ExtendableError {
+class ArgumentError extends ExtendableError {
   /**
-   * Creates a new InvalidArgumentError.
-   * @param  {string} message Error message.
+   * Creates a new ArgumentError.
+   *
+   * Note, that there are error codes:
+   * - 1 Missing argument.
+   * - 2 Wrong type of argument.
+   * - 3 Invalid argument.
+   *
+   * @param  {string}   message   Error message.
+   * @param  {integer}  code      Error code.
    */
-  constructor(message) {
+  constructor(message, code) {
+    // Set message.
     super(message);
+
+    // Set code.
+    this.code = code;
   }
 }
 
-module.exports = InvalidArgumentError;
+module.exports = ArgumentError;
