+++ lib/error/RequestError.js
@@ -1,11 +1,11 @@
 /**
- * @name Request error.
+ * @name Request error
  *
  * @description Error class for errors while requesting data.
  *
  * @example
  * 	const RequestError = require('./RequestError');
- * 	throw new RequestError('Expected ID');
+ * 	throw new RequestError('Could not resolve IP', { a: 1 });
  */
 
 'use strict';
@@ -19,10 +19,16 @@ const ExtendableError = require('./ExtendableError');
 class RequestError extends ExtendableError {
   /**
    * Creates a new RequestError.
-   * @param  {string} message Error message.
+   *
+   * @param {string} message  Error message.
+   * @param {object} data     Error data.
    */
-  constructor(message) {
+  constructor(message, data) {
+    // Set message.
     super(message);
+
+    // Set data.
+    this.data = data;
   }
 }
