+++ lib/provider/constants.js
@@ -1,0 +1,13 @@
+/**
+ * @name Constants provider.
+ *
+ * @description Contains constants, mostly used for codes.
+ *
+ * @type {Object}
+ */
+module.exports = {
+  // ArgumentError.
+  ARGUMENT_MISSING: 1,
+  ARGUMENT_WRONG_TYPE: 2,
+  ARGUMENT_INVALID: 3,
+};
